refactor: 디렉토리 구조 변경
범용성이 높은 컴포넌트는 common에,
이 어플리케이션에서만 사용하는 컴포넌트는 domain에 배치
